+++ spec/rubocop/cop/style/rbs_inline/data_define_with_block_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::Style::RbsInline::DataDefineWithBlock, :config do
-  let(:config) { RuboCop::Config.new }
+  let(:config) { rbs_inline_config }
 
   context "when Data.define is called with a do...end block" do
     it "registers an offense" do