+++ spec/rubocop/cop/style/rbs_inline/invalid_types_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe RuboCop::Cop::Style::RbsInline::InvalidTypes, :config do
-  let(:config) { RuboCop::Config.new }
+  let(:config) { rbs_inline_config }
 
   context "when using `#bad_method`" do
     it "registers an offense" do