+++ spec/rubocop/cop/style/rbs_inline/redundant_type_annotation_spec.rb
@@ -4,6 +4,7 @@ RSpec.describe RuboCop::Cop::Style::RbsInline::RedundantTypeAnnotation, :config 
   context "when EnforcedStyle is method_type_signature" do
     let(:config) do
       RuboCop::Config.new(
+        "Style/RbsInline" => { "Mode" => "opt_out" },
         "Style/RbsInline/RedundantTypeAnnotation" => {
           "EnforcedStyle" => "method_type_signature",
           "SupportedStyles" => %w[method_type_signature doc_style doc_style_and_return_annotation]
@@ -269,6 +270,7 @@ RSpec.describe RuboCop::Cop::Style::RbsInline::RedundantTypeAnnotation, :config 
   context "when EnforcedStyle is doc_style" do
     let(:config) do
       RuboCop::Config.new(
+        "Style/RbsInline" => { "Mode" => "opt_out" },
         "Style/RbsInline/RedundantTypeAnnotation" => {
           "EnforcedStyle" => "doc_style",
           "SupportedStyles" => %w[method_type_signature doc_style doc_style_and_return_annotation]
@@ -536,6 +538,7 @@ RSpec.describe RuboCop::Cop::Style::RbsInline::RedundantTypeAnnotation, :config 
   context "when EnforcedStyle is doc_style_and_return_annotation" do
     let(:config) do
       RuboCop::Config.new(
+        "Style/RbsInline" => { "Mode" => "opt_out" },
         "Style/RbsInline/RedundantTypeAnnotation" => {
           "EnforcedStyle" => "doc_style_and_return_annotation",
           "SupportedStyles" => %w[method_type_signature doc_style doc_style_and_return_annotation]